requirements.txt
# Core
fastapi==0.111.0
uvicorn[standard]==0.29.0

# Database
psycopg2-binary==2.9.9

# Password hashing
bcrypt==4.1.3
passlib[bcrypt]==1.7.4

# HTTP client (TMDB API)
requests==2.32.3

# Environment variable loading
python-dotenv==1.0.1

# Data validation (FastAPI models — used from Milestone 2 onward)
pydantic==2.7.1
pydantic-settings==2.2.1
